Add comprehensive tests for animation resampling, provenance handling, and FBX timebase
- Introduced tests for resampling animation scenes with varying frame rates and ensuring correct endpoint values.
- Added provenance tests to validate deterministic behavior and handle malformed data gracefully.
- Implemented tests for encoding contracts to ensure proper text handling across different locales.
- Created tests for FBX declared timebase resolution and compatibility classification for optional hierarchy.
- Included tests for project timing migration to ensure legacy sampling modes are preserved and validated correctly.

+++ dlanm2_gui/root_motion_override.py
@@ -30,7 +30,7 @@ except ImportError:
 MOTION_HELPER_DESCRIPTOR = 0xCCC3CDDF
 
 
-def _frame_ticks(document: FbxDocument, fps: int, frame_count: int) -> list[int]:
+def _frame_ticks(document: FbxDocument, fps: float, frame_count: int) -> list[int]:
     if hasattr(document, "frame_ticks"):
         ticks = list(document.frame_ticks(fps=fps))
     else:
@@ -71,7 +71,7 @@ def apply_root_motion_source_override(
     canonical_smd: str | Path,
     source_bone: str,
     root_policy: str,
-    fps: int,
+    fps: float,
     animation_stack: str | None = None,
     source_bone_aliases: Mapping[str, str] | None = None,
     target_root_bone: str = "bip01",
